pyproject.toml
[project]
name = "fastapi-server-template"
version = "0.1.0"
description = "A simple FastAPI server template"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
]

